+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 1.14.0+2 (not released yet)
+* API change: Changed `submittable` type from `bool` to `SubmittableMeta` in the `ProvidedMeta`.
 * Added an `activatable` flag to an action.
 
 ## 1.14.0+1 (2019-12-21)